Updating import scripts
